+++ Shore/Shore/CodeAnalysis/Syntax/Lexer.cs
@@ -247,7 +247,7 @@ namespace Shore.CodeAnalysis.Syntax
 
             var length = _position - _start;
             var text = _text.ToString(_start, length);
-            if (!float.TryParse(text, out var value))
+            if (!double.TryParse(text, out var value))
             {
                 var span = new TextSpan(_start, length);
                 var location = new TextLocation(_text, span);
